Fix depth of field blur not applying to squares
This happened because the variable for blur references the size of the square. When that size variable is defined in the square's inline styling, that variable is only available inside of the square CSS selector. When the variable for blur was defined in the document root, this did not work.